refactor to include indexer
