+++ pkg/uevent/listener.go
@@ -51,7 +51,15 @@ var (
 
 	errEmptyBuf  = errors.New("buffer is empty")
 	errShortRead = errors.New("short read")
+
+	resyncPeriod = 60 * time.Second
 )
+
+type DeviceUEventHandler interface {
+	Add(context.Context, *sys.Device) error
+	Change(context.Context, *sys.Device) error
+	Remove(context.Context, *sys.Device) error
+}
 
 type listener struct {
 	sockfd      int
@@ -60,6 +68,8 @@ type listener struct {
 
 	nodeID  string
 	handler DeviceUEventHandler
+
+	indexer *indexer
 }
 
 type deviceEvent struct {
@@ -71,6 +81,182 @@ type deviceEvent struct {
 	timer   *time.Timer
 
 	udevData *sys.UDevData
+}
+
+func Run(ctx context.Context, nodeID string, handler DeviceUEventHandler) error {
+	sockfd, err := syscall.Socket(
+		syscall.AF_NETLINK,
+		syscall.SOCK_RAW,
+		syscall.NETLINK_KOBJECT_UEVENT,
+	)
+	if err != nil {
+		return err
+	}
+
+	if err := syscall.Bind(sockfd, &syscall.SockaddrNetlink{
+		Family: syscall.AF_NETLINK,
+		Pid:    uint32(os.Getpid()),
+		Groups: 2,
+	}); err != nil {
+		return err
+	}
+
+	listener := &listener{
+		sockfd:     sockfd,
+		handler:    handler,
+		eventQueue: newEventQueue(),
+		nodeID:     nodeID,
+		indexer:    newIndexer(ctx, resyncPeriod),
+	}
+
+	go listener.processEvents(ctx)
+
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+			dEvent, err := listener.getNextDeviceUEvent(ctx)
+			if err != nil {
+				return err
+			}
+			listener.eventQueue.push(dEvent)
+		}
+	}
+}
+
+func (l *listener) processEvents(ctx context.Context) {
+	for {
+		select {
+		case <-ctx.Done():
+			klog.Error(ctx.Err())
+			return
+		default:
+			dEvent := l.eventQueue.pop()
+			if err := dEvent.collectUDevData(); err != nil {
+				klog.ErrorS(err, "failed to collect udevdata for path: %s", dEvent.devPath)
+				l.eventQueue.push(dEvent)
+				continue
+			}
+			if err := l.handle(ctx, dEvent); err != nil {
+				klog.ErrorS(err, "failed to handle an event", dEvent.action)
+				// Push it again to the queue
+				l.eventQueue.push(dEvent)
+			}
+		}
+	}
+}
+
+func (l *listener) handle(ctx context.Context, dEvent *deviceEvent) error {
+	if sys.IsLoopBackDevice(dEvent.udevData.Path) {
+		klog.V(5).InfoS(
+			"loopback device is ignored",
+			"ACTION", dEvent.action,
+			"DEVPATH", dEvent.devPath)
+		return nil
+	}
+
+	name := filepath.Base(udevData.Path)
+	if name == "" {
+		return fmt.Errorf("udevData does not have valid DEVPATH %v", udevData.Path)
+	}
+
+	device = &sys.Device{
+		Name:         name,
+		Major:        udevData.Major,
+		Minor:        udevData.Minor,
+		Virtual:      strings.Contains(udevData.Path, "/virtual/"),
+		Partition:    udevData.Partition,
+		WWID:         udevData.WWID,
+		Model:        udevData.Model,
+		UeventSerial: udevData.UeventSerial,
+		Vendor:       udevData.Vendor,
+		DMName:       udevData.DMName,
+		DMUUID:       udevData.DMUUID,
+		MDUUID:       udevData.MDUUID,
+		PTUUID:       udevData.PTUUID,
+		PTType:       udevData.PTUUID,
+		PartUUID:     udevData.PartUUID,
+		UeventFSUUID: udevData.UeventFSUUID,
+		FSType:       udevData.FSType,
+		FSUUID:       udevData.FSUUID,
+	}
+
+	if ok := listener.indexer.validateDevice(device); ok {
+		return nil
+	}
+
+	if err := device.ProbeHostInfo(); err != nil {
+		// if drive is deleted
+		if !errors.Is(fs.ErrNotExist()) {
+			return err
+		}
+	}
+
+	switch dEvent.action {
+	case Add:
+		return l.handler.Add(ctx, device)
+	case Change:
+		return l.handler.Change(ctx, device, listener.indexer.getDeviceCRD(device))
+	case Remove:
+		return l.handler.Remove(ctx, device, listener.indexer.getDeviceCRD(device))
+	default:
+		return fmt.Errorf("invalid device action: %s", dEvent.action)
+	}
+}
+
+func (l *listener) getNextDeviceUEvent(ctx context.Context) (*deviceEvent, error) {
+	for {
+		buf, err := l.readMsg()
+		if err != nil {
+			return nil, err
+		}
+
+		dEvent, err := l.parseUEvent(buf)
+		if err != nil {
+			if errors.Is(err, errNonDeviceEvent) {
+				continue
+			}
+			return nil, err
+		}
+		return dEvent, nil
+	}
+}
+
+func (dEvent *deviceEvent) collectUDevData() error {
+	switch dEvent.action {
+	case Add, Change:
+		// Older kernels like in CentOS 7 does not send all information about the device,
+		// hence read relevant data from /run/udev/data/b<major>:<minor>
+		runUdevDataMap, err := sys.ReadRunUdevDataFile(dEvent.udevData.Major, dEvent.udevData.Minor)
+		if err != nil {
+			return err
+		}
+		runUdevData, err := mapToUdevData(runUdevDataMap)
+		if err != nil {
+			return err
+		}
+		// Fill the missing fields
+		return dEvent.fillMissingUdevData(runUdevData)
+	case Remove:
+		// Removed device cannot be probed locally
+		// Relying on the event data
+		return nil
+	default:
+		return fmt.Errorf("invalid device action: %s", dEvent.action)
+	}
+}
+
+func (dEvent *deviceEvent) toDevice() (*sys.Device, error) {
+	switch dEvent.action {
+	case Add, Change:
+		return sys.CreateDevice(dEvent.udevData)
+	case Remove:
+		// Removed device cannot be probed locally
+		return sys.NewDevice(dEvent.udevData)
+	default:
+		return nil, fmt.Errorf("invalid device action: %s", dEvent.action)
+	}
 }
 
 func (dEvent *deviceEvent) fillMissingUdevData(runUdevData *sys.UDevData) error {
@@ -204,271 +390,4 @@ func (dEvent *deviceEvent) fillMissingUdevData(runUdevData *sys.UDevData) error 
 	}
 
 	return nil
-}
-
-func (dEvent *deviceEvent) collectUDevData() error {
-	switch dEvent.action {
-	case Add, Change:
-		// Older kernels like in CentOS 7 does not send all information about the device,
-		// hence read relevant data from /run/udev/data/b<major>:<minor>
-		runUdevDataMap, err := sys.ReadRunUdevDataFile(dEvent.udevData.Major, dEvent.udevData.Minor)
-		if err != nil {
-			return err
-		}
-		runUdevData, err := mapToUdevData(runUdevDataMap)
-		if err != nil {
-			return err
-		}
-		// Fill the missing fields
-		return dEvent.fillMissingUdevData(runUdevData)
-	case Remove:
-		// Removed device cannot be probed locally
-		// Relying on the event data
-		return nil
-	default:
-		return fmt.Errorf("invalid device action: %s", dEvent.action)
-	}
-}
-
-func (dEvent *deviceEvent) toDevice() (*sys.Device, error) {
-	switch dEvent.action {
-	case Add, Change:
-		return sys.CreateDevice(dEvent.udevData)
-	case Remove:
-		// Removed device cannot be probed locally
-		return sys.NewDevice(dEvent.udevData)
-	default:
-		return nil, fmt.Errorf("invalid device action: %s", dEvent.action)
-	}
-}
-
-type DeviceUEventHandler interface {
-	Add(context.Context, *sys.Device) error
-	Change(context.Context, *sys.Device) error
-	Remove(context.Context, *sys.Device) error
-}
-
-func Run(ctx context.Context, nodeID string, handler DeviceUEventHandler) error {
-	sockfd, err := syscall.Socket(
-		syscall.AF_NETLINK,
-		syscall.SOCK_RAW,
-		syscall.NETLINK_KOBJECT_UEVENT,
-	)
-	if err != nil {
-		return err
-	}
-
-	if err := syscall.Bind(sockfd, &syscall.SockaddrNetlink{
-		Family: syscall.AF_NETLINK,
-		Pid:    uint32(os.Getpid()),
-		Groups: 2,
-	}); err != nil {
-		return err
-	}
-
-	listener := &listener{
-		sockfd:     sockfd,
-		handler:    handler,
-		eventQueue: newEventQueue(),
-		nodeID:     nodeID,
-	}
-
-	listener.processEvents(ctx)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-			dEvent, err := listener.getNextDeviceUEvent(ctx)
-			if err != nil {
-				return err
-			}
-			listener.eventQueue.push(dEvent)
-		}
-	}
-}
-
-func (l *listener) handle(ctx context.Context, dEvent *deviceEvent) error {
-	if sys.IsLoopBackDevice(dEvent.udevData.Path) {
-		klog.V(5).InfoS(
-			"loopback device is ignored",
-			"ACTION", dEvent.action,
-			"DEVPATH", dEvent.devPath)
-		return nil
-	}
-
-	device, err := dEvent.toDevice()
-	if err != nil {
-		klog.ErrorS(err, "ACTION", dEvent.action, "DEVPATH", dEvent.devPath)
-		return err
-	}
-
-	switch dEvent.action {
-	case Add:
-		return l.handler.Add(ctx, device)
-	case Change:
-		return l.handler.Change(ctx, device)
-	case Remove:
-		return l.handler.Remove(ctx, device)
-	default:
-		return fmt.Errorf("invalid device action: %s", dEvent.action)
-	}
-}
-
-func (l *listener) processEvents(ctx context.Context) {
-	go func() {
-		for {
-			select {
-			case <-ctx.Done():
-				klog.Error(ctx.Err())
-				return
-			default:
-				dEvent := l.eventQueue.pop()
-				if err := dEvent.collectUDevData(); err != nil {
-					klog.ErrorS(err, "failed to collect udevdata for path: %s", dEvent.devPath)
-					l.eventQueue.push(dEvent)
-					continue
-				}
-				if err := l.handle(ctx, dEvent); err != nil {
-					klog.ErrorS(err, "failed to handle an event", dEvent.action)
-					// Push it again to the queue
-					l.eventQueue.push(dEvent)
-				}
-			}
-		}
-	}()
-}
-
-func (l *listener) getNextDeviceUEvent(ctx context.Context) (*deviceEvent, error) {
-	for {
-		buf, err := l.ReadMsg()
-		if err != nil {
-			return nil, err
-		}
-
-		dEvent, err := l.parseUEvent(buf)
-		if err != nil {
-			if errors.Is(err, errNonDeviceEvent) {
-				continue
-			}
-			return nil, err
-		}
-		return dEvent, nil
-	}
-}
-
-func parse(msg []byte) (map[string]string, error) {
-	if !bytes.HasPrefix(msg, []byte(libudev)) {
-		return nil, errors.New("libudev signature not found")
-	}
-
-	// magic number is stored in network byte order.
-	if magic := binary.BigEndian.Uint32(msg[8:]); magic != libudevMagic {
-		return nil, fmt.Errorf("libudev magic mismatch; expected: %v, got: %v", libudevMagic, magic)
-	}
-
-	offset := int(msg[16])
-	if offset < 17 {
-		return nil, fmt.Errorf("payload offset %v is not more than 17", offset)
-	}
-	if offset > len(msg) {
-		return nil, fmt.Errorf("payload offset %v beyond message length %v", offset, len(msg))
-	}
-
-	fields := bytes.Split(msg[offset:], fieldDelimiter)
-	event := map[string]string{}
-	for _, field := range fields {
-		if len(field) == 0 {
-			continue
-		}
-		switch tokens := strings.SplitN(string(field), "=", 2); len(tokens) {
-		case 1:
-			event[tokens[0]] = ""
-		case 2:
-			event[tokens[0]] = tokens[1]
-		}
-	}
-	return event, nil
-}
-
-func (l *listener) parseUEvent(buf []byte) (*deviceEvent, error) {
-	eventMap, err := parse(buf)
-	if err != nil {
-		return nil, err
-	}
-
-	if eventMap["SUBSYSTEM"] != "block" {
-		return nil, errNonDeviceEvent
-	}
-
-	eventAction := action(eventMap["ACTION"])
-	switch eventAction {
-	case Add, Change, Remove:
-	default:
-		return nil, fmt.Errorf("invalid device action: %s", eventAction)
-	}
-
-	udevData, err := mapToUdevData(eventMap)
-	if err != nil {
-		return nil, err
-	}
-
-	return &deviceEvent{
-		created:  time.Now().UTC(),
-		action:   eventAction,
-		udevData: udevData,
-		devPath:  udevData.Path,
-	}, nil
-}
-
-func (l *listener) msgPeek() (int, []byte, error) {
-	var n int
-	var err error
-	buf := make([]byte, os.Getpagesize())
-	for {
-		if n, _, err = syscall.Recvfrom(l.sockfd, buf, syscall.MSG_PEEK); err != nil {
-			return n, nil, err
-		}
-
-		if n < len(buf) {
-			break
-		}
-
-		buf = make([]byte, len(buf)+os.Getpagesize())
-	}
-
-	buf = buf[:n]
-
-	return n, buf, err
-}
-
-func (l *listener) msgRead(buf []byte) error {
-	if buf == nil {
-		return errEmptyBuf
-	}
-
-	n, _, err := syscall.Recvfrom(l.sockfd, buf, 0)
-	if err != nil {
-		return err
-	}
-
-	if n != len(buf) {
-		return errShortRead
-	}
-
-	return nil
-}
-
-// ReadMsg allow to read an entire uevent msg
-func (l *listener) ReadMsg() ([]byte, error) {
-	_, buf, err := l.msgPeek()
-	if err != nil {
-		return nil, err
-	}
-	if err = l.msgRead(buf); err != nil {
-		return nil, err
-	}
-
-	return buf, nil
 }
